Nonfunctioning implementation of Chan's algorithm
Implementation of Chan's algorithm, with known errors. This algorithm fails to compute the correct hull, because of issues in when rtangent is called.
The calls to rtangent fail to always fullfil the preconditions, and sometimes p is colinear with all points in v, so the output is incorrect. I have
an idea on how to resolve it, but it will require refactoring the algorithm, so pushing the current version to source control