Refactor chan's to use subhull indices
Begin fixing the errors in Chan's algorithm, using subhull indices rather than the points themselves, allowing us to handle the case where the most recent point is in the hull we're searching, without using rtangent.
This is important because the rtangent preconditions require that the point be exterior to the hull v, and when p is on v, that fails. Several other failure cases remain, particularly where p is colinear with all points
in a subhull v, and when a partition has size 1.