Add Synthetic Datasets
Add a set of synthetic datasets, going from size 10 to 100k for a
dataset that is the worst case for Andrew's algorithm, points on the
unit disk, points spaced evenly around the unit circle, and points on
the unit disk with three points well outside that make h = 3. The points
on the unit disk and the points on the unit disk with points outside to
make h = 3 also have a version with 1 million points.

The datasets containing worst case data for quickhull only exist for 10
and 30 points, due to limitations on floating point precision.